+++ docs/changes.rst
@@ -6,6 +6,7 @@ Version 0.1.1
 
 To be released.
 
+- Added :class:`sider.types.Boolean` type.
 - Added :class:`sider.types.Date` type.
 - Added :class:`sider.datetime.FixedOffset` tzinfo subtype.
 - Added :class:`sider.types.DateTime` and